% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_png.R
\name{custom_png}
\alias{custom_png}
\title{custom_png
modification of png function from grDevices}
\usage{
custom_png(
  filename = "Rplot\%03d.png",
  width = 1200,
  height = 700,
  units = "px",
  pointsize = 12,
  bg = "white",
  res = NA,
  family = "",
  restoreConsole = TRUE,
  type = "cairo",
  antialias = c("default", "none", "cleartype", "gray", "subpixel"),
  symbolfamily = "default"
)
}
\description{
default type is "Cairo"
default width and height is 1200×700 px
for more details see function png {grDevices}
}
\examples{
png(filename = "Rplot\%03d.png",
width = 1200, height = 700, units = "px", pointsize = 12,
bg = "white", res = NA, family = "", restoreConsole = TRUE,
type = "cairo", antialias, symbolfamily="default")
}
