% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_ASITIS.R
\name{get_colors_ASITIS}
\alias{get_colors_ASITIS}
\title{Colors for ASITIS company visual style for data visualization}
\usage{
get_colors_ASITIS(n = NA, dark = F)
}
\arguments{
\item{n}{Requested count of colors in color vector, if not provided, returns complete list of available colors}

\item{dark}{only is used if requested color count is specified, if TRUE, colors will be selected from pallet "Osmibarevná paleta tmavá" instead od "Osmibarevná paleta světlá"}
}
\value{
list of vectors of HEX color codes or one vector if number of colors is specified
}
\description{
Function returns complete list of 6 solitaire colors and 11 color vectors (palettes) or one color vector of given color number
}
\examples{
get_colors_ASITIS();
get_colors_ASITIS(8, dark = TRUE);
}
