% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_ASITIS.R
\name{get_colors_ASITIS_marketing}
\alias{get_colors_ASITIS_marketing}
\title{Colors for ASITIS company visual style for marketing and general use}
\usage{
get_colors_ASITIS_marketing()
}
\value{
list of vectors of HEX color codes
}
\description{
Function returns complete list of 2 solitaire colors and 7 color vectors (palettes)
}
\examples{
get_colors_ASITIS_marketing();
}
