% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_ASITIS.R
\name{get_colours_ASITIS}
\alias{get_colours_ASITIS}
\title{Colours for ASITIS company visual style for data visualization}
\usage{
get_colours_ASITIS(n = NA, dark = F)
}
\arguments{
\item{n}{Requested number of colours in colour vector, if not provided, returns complete list of available colours}

\item{dark}{only is used if requested colour count is specified, if TRUE, colours will be selected from pallet "Osmibarevná paleta tmavá" instead od "Osmibarevná paleta světlá"}
}
\value{
list of vectors of HEX color codes or one vector if number of colors is specified
}
\description{
Function returns complete list of 6 solitaire colours and 11 colour vectors (palletes) or one colour vector of given colour number
}
\examples{
get_colors_ASITIS();
get_colors_ASITIS(8, dark = TRUE);
}
