% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_ASITIS.R
\name{get_colours_ASITIS_marketing}
\alias{get_colours_ASITIS_marketing}
\title{Colours for ASITIS company visual style for marketing and general use}
\usage{
get_colours_ASITIS_marketing()
}
\value{
list of vectors of HEX color codes
}
\description{
Function returns complete list of 2 solitaire colours and 7 colour vectors (palletes)
}
\examples{
get_colors_ASITIS_marketing();
}
