% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ASITIS.R
\name{theme_ASITIS}
\alias{theme_ASITIS}
\title{ASITIS company ggplot theme
version 1.2.1}
\usage{
theme_ASITIS(
  angle.x = FALSE,
  title.on.center = FALSE,
  no.grid = FALSE,
  double.axis = FALSE,
  double.axis.reverse = FALSE
)
}
\arguments{
\item{angle.x}{boolean, if true, values on x axis will be 15° slanted}

\item{title.on.center}{boolean, if true, plot title will be aligned to center}

\item{no.grid}{boolean, if true, plot will not have visible grid}

\item{double.axis}{boolean, if true, plot will have both right (red colored) and left (blue colored) y axes}

\item{double.axis.reverse}{boolean, if true, plot will have both right (blue colored) and left (red colored) y axes}
}
\value{
ggplot theme
}
\description{
based on ggplot2 theme_bw
}
\examples{
p + theme_ASITIS();
p + theme_ASITIS(angle.x = T, title.on.center = T);
}
