% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ASITIS_old.R
\name{theme_ASITIS_old}
\alias{theme_ASITIS_old}
\title{ASITIS company ggplot theme
version 1.2.1 old style modification}
\usage{
theme_ASITIS_old(
  angle.x = FALSE,
  title.on.center = FALSE,
  no.grid = FALSE,
  double.axis = FALSE,
  double.axis.reverse = FALSE
)
}
\arguments{
\item{angle.x}{boolean, if true, values on x axis will be 15° slanted}

\item{title.on.center}{boolean, if true, plot title will be aligned to the center}

\item{no.grid}{boolean, if true, plot will not have visible grid}

\item{double.axis}{boolean, if true, plot will have both right (red colored) and left (blue colored) y axes}

\item{double.axis.reverse}{boolean, if true, plot will have both right (blue colored) and left (red colored) y axes}
}
\value{
ggplot theme
}
\description{
Modification of standard style theme_ASITIS to replicate old plot style used before summer 2024.
Use with standard print function, not with printA intended for use with theme_ASITIS.
}
\examples{
p + theme_ASITIS();
p + theme_ASITIS(angle.x = T, title.on.center = T);
}
